Ensure crash in commit/rollback does not try twice
It is possible that if during the rollback or commit command
that there may be a disk error that will cause SQLite to return
an error.  When that happens we must not attempt to rollback again.

This was seen when running multi-threaded and a bug in some code
resulted in the SQLite file descripter being closed in the middle of
a transaction.